% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionBNGdataprep.R
\name{bngprep}
\alias{bngprep}
\title{Prepare data from NBN gateway or atlas for use in species distribution modelling.}
\usage{
bngprep(speciesdf, precisionCol = "precision", bngCol = "gridReference",
  datafrom = NULL, minyear, mindata = 5000, covarRes)
}
\arguments{
\item{speciesdf}{Data frame exported from NBN gateway or NBN atlas with data for a species.}

\item{precisionCol}{The column name of the column in \code{speciesdf} denoting the precision of the species record locations.}

\item{bngCol}{The column name of the column in\code{speciesdf} giving the species record location as a BNG grid reference.}

\item{datafrom}{Character, one of "NBNgateway" or "NBNatlas", indicating the data source.}

\item{minyear}{Numeric, the earliest year from which data should be used. Data older than this will be discarded.}

\item{mindata}{The target minimum number of data points to return. If this is specified, the lowest resolution data will be discarded if there are enough higher resolution data points available to reach this target.}

\item{covarRes}{The resolution of the environmental covariate data layers, in metres. Data will not be discarded if it is of higher resolution than the environmental covariate layers.}
}
\value{
A copy of \code{speciesdf} with absence data removed, easting and northing columns generated from BNG grid references and data from Northern Ireland removed.
}
\description{
This function removes absence data (as the models generate their own pseudo-absences), converts British National Grid grid references into easting and northing coordinates and removes data from Northern Ireland to allow modelling using GB data layers.
}
