% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionMultimod.R
\name{Multi_mod}
\alias{Multi_mod}
\title{Set up modelling multiple species from NBN gateway or atlas for use in species distribution modelling.}
\usage{
Multi_mod(sp_list = sp_list, out_flder = "Outputs/",
  dat_flder = "Inputs/", bkgd_flder = "BGmasks/", vars, max_tries = 1,
  datafrom = "NBNgatweay", minyear = 0, maxyear = 0, mindata = 5000,
  covarRes = 300, models = c("MaxEnt", "BioClim", "SVM", "RF", "GLM", "GAM",
  "BRT"), prop_test_data = 0.25, bngCol = "gridReference",
  mult_prssr = FALSE, rndm_occ = TRUE)
}
\arguments{
\item{sp_list}{List of unique species names which you wish to model.}

\item{out_flder}{The location of the output folder for your models.}

\item{dat_flder}{The location of the folder containing your species occurrence data, as data frames exported from NBN gateway or NBN atlas with files containing data for a single species.}

\item{bkgd_flder}{The location of the folder containing your background masks. These should be raster files showing the background area in which pseudo-absence points will be placed. Cells from which background points should be taken should have a value of 1 and excluded cells should be NA.}

\item{vars}{A RasterStack of the environmental parameters to be used as predictor variables for the species range.}

\item{max_tries}{The number of times the number is run.}

\item{datafrom}{Whether it is data from the "NBNgateway" or "NBNatlas".}

\item{minyear}{Numeric, the earliest year from which data should be selected. Year inclusive, data older than this will be discarded.}

\item{maxyear}{Numeric, the latest year from which data should be used. Year inclusive, data newer than this will be discarded.}

\item{mindata}{The target minimum number of data points to return. If this is specified, the lowest resolution data will be discarded if there are enough higher resolution data points available to reach this target.}

\item{covarRes}{The resolution of the environmental covariate data layers, in metres. Data will not be discarded if it is of higher resolution than the environmental covariate layers.}

\item{models}{A character vector of the models to run and evaluate. This should be at least one of \code{"MaxEnt"}, \code{"BioClim"}, \code{"SVM"}, \code{"RF"}, \code{"GLM"}, \code{"GAM"}, \code{"BRT"}. Default is to run all models.}

\item{prop_test_data}{Numeric, the proportion of data to keep back as testing data for evaluating the models. Default is 25\%.}

\item{bngCol}{The column name of the column in \code{speciesdf} giving the species record location as a BNG grid reference. For NBNatlas, this will vary with recorder precision, so you should select the most appropriate for your data, for example "OSGR 1km", "OSGR 2km" or "OSGR 10km".}

\item{mult_prssr}{Set up a parallel backend to use multiple processors. As a default this is turned off. Need to ensure the suggested packages have been loaded in order to run this.}

\item{rndm_occ}{Logical, Default is TRUE and will randomise the locations of presence points where the species occurrence data is low resolution, through calling the randomOcc function.}
}
\value{
A copy
}
\description{
This function is to carry out modelling of multiple species using the JNCCSdms package.
}
