% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionBNGdataprep.R
\name{bngprep}
\alias{bngprep}
\title{Prepare data from NBN gateway or atlas for use in species distribution modelling.}
\usage{
bngprep(speciesdf, bngCol = "OSGR", precisionCol = "precision",
  datafrom = "NBNatlas", minyear = 0, maxyear = 0, mindata = 5000,
  covarRes = 300)
}
\arguments{
\item{speciesdf}{Data frame exported from NBN gateway or NBN atlas with data for a species.}

\item{bngCol}{The column name of the column in \code{speciesdf} giving the species record location as a BNG grid reference. For NBNgateway this is the 'gridReference' field. The default is set to the 'OSGR' field which is used for data from NBNatlas and is also used to generate the location precision.}

\item{precisionCol}{The column name of the column in \code{speciesdf} denoting the precision of the species record locations. This only needs to be specified for NBNgateway data as this is generated from the selected \code{bngCol} for data from the NBNatlas.}

\item{datafrom}{Character, one of 'NBNgateway' or 'NBNatlas', indicating the data source.'NBNatlas' is the default.The format of data from the NBN Gateway and NBN Atlas differ, which is why the 'datafrom' field must be specified when preparing the data using the bngprep function.}

\item{minyear}{Numeric, the earliest year from which data should be selected. Year inclusive, data older than this will be discarded.}

\item{maxyear}{Numeric, the latest year from which data should be used. Year inclusive, data newer than this will be discarded.}

\item{mindata}{The target minimum number of data points to return. If this is specified, the lowest resolution data will be discarded if there are enough higher resolution data points available to reach this target.}

\item{covarRes}{The resolution of the environmental covariate data layers, in metres. Data will not be discarded if it is of higher resolution than the environmental covariate layers.}
}
\value{
A copy of \code{speciesdf} with absence data removed, easting and northing columns generated from BNG grid references and data from Northern Ireland removed.
}
\description{
This function removes absence data (as the models generate their own pseudo-absences), converts British National Grid grid references into easting and northing coordinates and removes data from Northern Ireland to allow modelling using GB data layers.
}
\examples{
#Examples using data from NBN Atlas:

data(ng_data)

bngprep(speciesdf = ng_data, bngCol = 'OSGR', datafrom = 'NBNatlas', mindata = 5000, minyear = 2007, covarRes = 300)
}
