% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionLatLondataprep.R
\name{latlonprep}
\alias{latlonprep}
\title{Prepare data from NBN Atlas for use in species distribution modelling.}
\usage{
latlonprep(speciesdf, xCol = "Latitude (WGS84)", yCol = "Longitude (WGS84)",
  precisionCol = "Coordinate uncertainty (m)", paCol = "Occurrence status",
  yearCol = "Start date year", minyear = 0, maxyear = 0, mindata = 500,
  covarRes = 300, GBonly = TRUE)
}
\arguments{
\item{speciesdf}{Data frame exported from the NBN atlas with data for a species.}

\item{xCol}{The column name of the column in \code{speciesdf} giving the species record location as a decimal latitude.}

\item{yCol}{The column name of the column in \code{speciesdf} giving the species record location as a decimal longitude.}

\item{precisionCol}{The column name of the column in \code{speciesdf} denoting the precision of the species record locations. For NBNAtlas this is denoted as the "Coordinate uncertainty (m)". where denoted with km or m, this will be converted into meters.}

\item{paCol}{The column name of the column in \code{speciesdf} denoting whether the species is present or absent.}

\item{yearCol}{The column name in \code{speciesdf} giving the year of the record.}

\item{minyear}{Numeric, the earliest year from which data should be selected. Year inclusive, data older than this will be discarded.}

\item{maxyear}{Numeric, the latest year from which data should be used. Year inclusive, data newer than this will be discarded.}

\item{mindata}{The target minimum number of data points to return. If this is specified, the lowest resolution data will be discarded if there are enough higher resolution data points available to reach this target.}

\item{covarRes}{The resolution of the environmental covariate data layers, in metres. Data will not be discarded if it is of higher resolution than the environmental covariate layers.
#'@param GBonly logical, TRUE if you wish to remove Northern Ireland from the records, FALSE if you wish to retain all records.}
}
\value{
A copy of \code{speciesdf} as a data frame with absence data removed, subset to your year range and area of interest.
}
\description{
This function removes absence data (as the models generate their own pseudo-absences), allows you to subset the data modelled based on year,
and allows modelling of GB and Northern Ireland using latitude and longitude coordinates.
}
\examples{
#test
}
