% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionOccurrence.R
\name{randomOcc}
\alias{randomOcc}
\title{Place occurence points at random within low-resolution grid cell}
\usage{
randomOcc(presframe, precisionCol = "precision")
}
\arguments{
\item{presframe}{Data frame of presence points. This should include columns titled "easting" and "northing", which should be x and y coordinates in metres (of the centre of the grid cell for gridded data), as well as a column giving the resolution (see \code{prescisionCol} below).}

\item{precisionCol}{The column indicating the resolution of the presence point data. This should be given as a grid cell size in metres e.g. for a 1km grid this should be 1000.}
}
\value{
A copy of the input dataframe \code{presframe}, but with new "easting" and "northing" values that place each point randomly within it's grid cell.
}
\description{
This function is used to randomise the location of a presence point where the species occurrence data is low resolution. For example, if your species occurrence data is at 1km resolution, but your environmental predictor variables are at 200m resolution, this enables you to randomise where within the 1km cell your species record is placed. The function can handle multiple resolutions within a single dataframe, as long as these are labelled with a precision column.
}
