% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionSDMs.R
\name{SDMs}
\alias{SDMs}
\title{Prepare presence and pseudo-absence data, run selected models and evaluate best model.}
\usage{
SDMs(occ = occurrence, bckg = background, varstack = vars,
  models = c("MaxEnt", "BioClim", "SVM", "RF", "GLM", "GAM", "BRT"),
  n_bg_points = nrow(pres_vars))
}
\arguments{
\item{occ}{A SpatialPointsDataFrame of presence points.}

\item{bckg}{A raster showing the background area in which pseudo-absence points will be placed. Cells from which background points should be taken should have a value of 1 and excluded cells should be NA.}

\item{varstack}{A RasterStack of the environmental parameters to be used as predictor variables for the species ramge.}

\item{models}{A character vector of the models to run and evaluate. This should be at least one of \code{"MaxEnt"}, \code{"BioClim"}, \code{"SVM"}, \code{"RF"}, \code{"GLM"}, \code{"GAM"}, \code{"BRT"}. Default is to run all models.}

\item{n_bg_points}{The number of pseudo-absence point to attempt to generate. Note that if a very restrictive mask is used the number actually generated may be fewer than that specified. Default is to attempt to generate the same number of pseudo-absences as presences for which there is data on the environmental parameters (this may be fewer than the number of points in \code{occ} if some of these fall in cells that are NA in one or more layers in \code{varstack}.}
}
\description{
This function uses the presence points
}
